finished sync button behavior [Fixed #77546948], finished adding transaction downloading from ziftr mock api and parsing mock json [Fixed #78843956]

+++ OneWallet/src/com/ziftr/android/ziftrwallet/sqlite/OWSQLiteOpenHelper.java
@@ -8,14 +8,11 @@ import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import com.ziftr.android.ziftrwallet.R;
 import com.ziftr.android.ziftrwallet.crypto.OWAddress;
 import com.ziftr.android.ziftrwallet.crypto.OWSha256Hash;
 import com.ziftr.android.ziftrwallet.crypto.OWTransaction;
 import com.ziftr.android.ziftrwallet.exceptions.OWAddressFormatException;
-import com.ziftr.android.ziftrwallet.fragment.accounts.OWWalletTransactionListAdapter;
 import com.ziftr.android.ziftrwallet.util.OWCoin;
-import com.ziftr.android.ziftrwallet.util.OWFiat;
 import com.ziftr.android.ziftrwallet.util.ZLog;
 
 /**
@@ -397,7 +394,7 @@ public class OWSQLiteOpenHelper extends SQLiteOpenHelper {
 	 */
 	public OWTransaction createTransaction(OWCoin coinId, BigInteger txAmount,
 			BigInteger txFee, List<String> displayAddresses, OWSha256Hash hash, 
-			String note, int numConfirmations) {
+			String note, long numConfirmations) {
 		OWTransaction tx = new OWTransaction(coinId, note, System.currentTimeMillis() / 100, txAmount);
 		tx.setSha256Hash(hash);
 		tx.setDisplayAddresses(displayAddresses);
